boskos: move handlers and tests into their own package
