+++ boskos/cmd/boskos/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 load("//prow:def.bzl", "prow_image")
 
 go_binary(
@@ -19,36 +19,18 @@ go_library(
     importpath = "k8s.io/test-infra/boskos/cmd/boskos",
     visibility = ["//visibility:private"],
     deps = [
-        "//boskos/common:go_default_library",
         "//boskos/crds:go_default_library",
+        "//boskos/handlers:go_default_library",
         "//boskos/ranch:go_default_library",
         "//prow/config:go_default_library",
         "//prow/interrupts:go_default_library",
         "//prow/logrusutil:go_default_library",
         "//prow/metrics:go_default_library",
         "//prow/pjutil:go_default_library",
-        "//prow/simplifypath:go_default_library",
         "@com_github_fsnotify_fsnotify//:go_default_library",
         "@com_github_prometheus_client_golang//prometheus:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_spf13_viper//:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = [
-        "boskos_test.go",
-        "server_client_test.go",
-        "storage_test.go",
-    ],
-    embed = [":go_default_library"],
-    deps = [
-        "//boskos/client:go_default_library",
-        "//boskos/common:go_default_library",
-        "//boskos/crds:go_default_library",
-        "//boskos/ranch:go_default_library",
-        "//boskos/storage:go_default_library",
     ],
 )
