+++ boskos/handlers/BUILD.bazel
@@ -2,19 +2,29 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["storage.go"],
-    importpath = "k8s.io/test-infra/boskos/storage",
+    srcs = ["handlers.go"],
+    importpath = "k8s.io/test-infra/boskos/handlers",
     visibility = ["//visibility:public"],
-    deps = ["//boskos/common:go_default_library"],
+    deps = [
+        "//boskos/common:go_default_library",
+        "//boskos/ranch:go_default_library",
+        "//prow/simplifypath:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+    ],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = ["storage_test.go"],
+    srcs = [
+        "handlers_test.go",
+        "server_client_test.go",
+    ],
     embed = [":go_default_library"],
     deps = [
+        "//boskos/client:go_default_library",
         "//boskos/common:go_default_library",
         "//boskos/crds:go_default_library",
+        "//boskos/ranch:go_default_library",
     ],
 )
